Added qunatme transformation
