+++ components/bpmn-q/modeler-component/extensions/quantme/QuantMEPlugin.js
@@ -9,6 +9,14 @@ import OpenToscaTab from "./configTabs/OpenToscaTab";
 import NisqAnalyzerTab from "./configTabs/NisqAnalyzerTab";
 import QrmDataTab from "./configTabs/QrmDataTab";
 import HybridRuntimeTab from "./configTabs/HybridRuntimeTab";
+import NotificationHandler from "../../editor/ui/notifications/NotificationHandler";
+import {getQRMs} from "./qrm-manager";
+import {startReplacementProcess} from "./replacement/QuantMETransformator";
+import {loadDiagram} from "../../common/util/IoUtilities";
+import {getModeler} from "../../editor/ModelerHandler";
+import * as camundaConfig from "../../editor/config/EditorConfigManager";
+import * as config from "./framework-config/config-manager";
+
 let quantMEModdleExtension = require('./resources/quantum4bpmn.json')
 
 export default {
@@ -43,11 +51,32 @@ export default {
     extensionModule: QuantMEExtensionModule,
     moddleDescription: quantMEModdleExtension,
     transformExtension: async () => {
-        // const modeler = getModeler();
-        // let xml = await getXml(modeler);
-        //
-        // await startReplacementProcess(xml, async function (xml) {
-        //     await saveXmlAsLocalFile(xml, "myProcess.bpmn");
-        // });
+        NotificationHandler.getInstance().displayNotification({
+            type: 'info',
+            title: 'Workflow Transformation Started!',
+            content: 'Successfully started transformation process for the current workflow!',
+            duration: 7000
+        });
+        const modeler = getModeler();
+
+        let xml = await modeler.get('bpmnjs').saveXML();
+        let currentQRMs = getQRMs();
+        let result = await startReplacementProcess(xml.xml, currentQRMs,
+            {
+                nisqAnalyzerEndpoint: config.getNisqAnalyzerEndpoint(),
+                transformationFrameworkEndpoint: config.getTransformationFrameworkEndpoint(),
+                camundaEndpoint: camundaConfig.getCamundaEndpoint()
+            });
+
+        if (result.status === 'transformed') {
+            await loadDiagram(result.xml, modeler);
+        } else {
+            NotificationHandler.getInstance().displayNotification({
+                type: 'warning',
+                title: 'Unable to transform workflow',
+                content: result.cause.toString(),
+                duration: 10000
+            });
+        }
     }
 }
